+++ NEWS.md
@@ -1,5 +1,8 @@
 # pkgdown (development version)
 
+* Missing topics makes the build fail when the environment variable `CI` is set
+  to `"true"` (@ThierryO, #1378).
+
 * The default navbar no longer includes a home icon - this took up precious
   horizontal space and wasn't very useful since there is already a link to the 
   home page immediately to its left (#1383).